Introduce clippy et al
This commit is preliminary to my main aim: introducing a finite source
of entropy from the Unstructured. I've gone through and taken clippy's
suggestions for the most part and believe that some of the code can be
redone with macros as a result. There are a good many changes
here. Uh, hopefully I can get all this merged upstream.

Still do to: documentation, macro fun times, introduce the finite
source.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>