Add a `depth` parameter to `size_hint` to avoid infinite recursion
Note that the implementations of `size_hint` only need to increment the depth at
indirection points where recursion can happen, e.g. at a `Box`. We take
advantage of this in the impls for `std` types, but conservatively do not take
advantage of that in the custom derive, since it doesn't have the necessary
info.

Fixes #30