Remove need for `synstructure` dependency
This dependency activates a number of features on `syn`, such as
`extra-traits`, which can cause excessive recompiles if `synstructure`
is part of some builds but not part of others. Especially for
`derive(Arbitrary)` it's often included in some builds (tests) but not
others (normal builds). By removing this dependency it should be a bit
speedier to build while also avoiding feature thrashing in normal builds.