Add the `Unstructured::arbitrary_loop` helper
This is a helper that I've defined and used a number of times in test case
generators. It executed the given function an arbitrary number of times. This is
useful when you want to generate N things but generating those things requires
some kind of context or bookkeeping so it isn't easy to just use `Vec<Thing> as
Arbitrary`.