Add `choose_index` -- a generalization of `choose`
`choose` comes only in one modality -- selecting a shared reference.
Often one wants an exclusive reference or maybe even to take ownership.

Rather than adding all those flavors, let's expose a function to
generate an index, the caller would then be able to DIY the rest.

Of course, the caller can open-code `u.gen_int(0..=array.len() - 1)`
themselves, but `-1` is annoying, and that'll also panic if the array
turns out to be empty. With `choose_index`, we can re-use the
`EmptyChoice` error.