fix(Unstructured)!: don't produce meaningless data if exhausted
If the Unstructured is completely exhausted, returning a buffer
completely filled with zeroes instead of indicating failure does not
make sense. It also causes issues for any users of Arbitrary
implementations for integers (which are based on this method) who rely
on eventually getting a nonzero result to e.g. break recursion.