Fix a bug where `int_in_range` could return out-of-range ints
We add one to the divisor for our modulo operation because we have an inclusive
range. But this could overflow when the range was already the full range of the
integer type. Now, we specifically check for that condition since we can
completely avoid the modulo operation in that case.

I also refactored the algorithm to avoid needing upcasts to `Int::Widest` since
we completely avoid overflows now, renamed some variables, and added some
comments to make everything a little easier to understand. Finally, I added a
test to ensure that we actually generate every integer in our given range.

Fixes #106