Require exact version of `arbitrary-derive`
`Arbitrary` had a new `try_size_hint` method added in #185 - while this
was semver safe for `arbitrary` because a default implementation was
provided which forwarded along, the new version of `arbitrary-derive`
which follows the same versioning as `arbitrary` started generating
`try_size_hint` methods, which causes v1.2.3 to fail to compile, as it
tries to grab `arbitrary-derive` "1.2.3" from crates.io, which gets
responded to with a "1.4.0" version.

Given that:
1. It is reasonable to evolve your traits this way, and it doesn't break
   semver on the base package
2. The patchlevel is specified on `arbitrary-derive`, so it seems the
   intent is to require a matching version
3. You likely want to keep the same version for your proc-macro as your
   main crate

This proposes using exact equality in the `Cargo.toml` so that future
trait updates of this nature won't break the ability to build older
packages via Cargo.