Replace `(usize, Option<usize>)` with custom type `SizeHint` and eliminate `try_size_hint()`.
The `SizeHint` type provides these advantages:

* Easier to construct by composition, using `&` and `|` or methods.
* More self-explanatory.
* Distinguishes between “unknown size” and “known to be unbounded”
  to aid debugging.

Since this is a breaking change to the trait, this commit also makes the
closely related change of combining `try_size_hint()` and `size_hint()`
into a single fallible function. This will make it easier to correctly
implement the trait.

Also, while I was touching the macro code, I made all the crate paths
absolute (`::arbitrary`), as they should be for the maximum compatibility.